Finish MCP server tool decomposition
